Exercise 1.1 part c
- Add a method to create a Tree from a string

Treat it as keys, with the values being the position in the string.
Taken from Sedgewick [1]

- Add a method to calculate the height of a Tree

Using this as a high-level indicator for whether a Tree is balanced. We
can see that the Tree from "abcdefghi" is not currently balanced, since
it has a height of 9. That means that it is very one-sided.

[1] https://www.amazon.co.uk/gp/product/B004P8J1NA/