Exercise 1.1 part d
Based on my research of Sedgewick [1] and Okasaki [2], I think a
red-black tree is a good data structure for a balanced search tree for
symbol tables.

For extra credit, I implemented a persistent red-black tree, based on my
reading of the Haskell version [3] (can’t quite understand the Standard
ML one yet).

This change shows that the height of the tree is reduced to 4 in both
previously pathological cases.

[1] https://www.amazon.co.uk/gp/product/B004P8J1NA/
[2] https://www.amazon.co.uk/gp/product/B00AKE1V04/
[3] https://wiki.rice.edu/confluence/download/attachments/2761212/Okasaki-Red-Black.pdf