Generify the Tree
An exercise in fighting the Java compiler. Done purely to try to
understand a bit more about Java’s type system, and the trade-offs that
were made when retrofitting generics to a language without breaking
backwards compatibility.

Also added some documentation. I’ve not found m(any?) examples of a
persistent red-black tree in Java, so there is some novelty to this
work.